feat(codegen): introduce target file in jump module to enhance read ability
